Sign binaries individually

+++ scripts/sign.sh
@@ -3,36 +3,37 @@
 INPUT=$1
 ENTITLEMENTS=$2
 
-# zip app
-chmod -R a-st $INPUT
-zip -r -q --symlinks unsigned.zip $INPUT
-rm -rf $INPUT
+FILE="$INPUT"
+NEEDS_ZIP=false
 
-# ensure storage server is clean
-ssh -q genie.theia@projects-storage.eclipse.org rm -f unsigned.zip signed.zip entitlements.plist
+# if folder, zip it
+if [ -d "$INPUT" ]; then
+    FILE=unsigned.zip
+    NEEDS_ZIP=true
 
-# copy zip to storage server
-scp unsigned.zip genie.theia@projects-storage.eclipse.org:./unsigned.zip
+    chmod -R a-st $INPUT
+    zip -r -q --symlinks $FILE $INPUT
+    rm -rf $INPUT
+fi
+
+# copy file to storage server
+scp $FILE genie.theia@projects-storage.eclipse.org:./
+rm -f $FILE
 
 # copy entitlements to storage server
 scp $ENTITLEMENTS genie.theia@projects-storage.eclipse.org:./entitlements.plist
 
 # sign over ssh
-ssh -q genie.theia@projects-storage.eclipse.org curl -o signed.zip -F file=@unsigned.zip -F entitlements=@entitlements.plist http://build.eclipse.org:31338/macsign.php
+ssh -q genie.theia@projects-storage.eclipse.org curl -o signed-$FILE -F file=@$FILE -F entitlements=@entitlements.plist http://build.eclipse.org:31338/macsign.php
 
-# copy signed app back from server
-scp genie.theia@projects-storage.eclipse.org:./signed.zip ./signed.zip
+# copy signed file back from server
+scp genie.theia@projects-storage.eclipse.org:./signed-$FILE $FILE
 
-# check for size
-actualSize=$(stat -f%z signed.zip)
+# ensure storage server is clean
+ssh -q genie.theia@projects-storage.eclipse.org rm -f $FILE signed-$FILE entitlements.plist
 
-# file is too small, it's likely an error so show it
-if [ $actualSize -lt 40000000 ]; then
-    echo "signed.zip is just $actualSize bytes large!"
-    output=$(cat signed.zip)
-    echo "$output"
+# if unzip needed
+if [ "$NEEDS_ZIP" = true ]; then
+    unzip -qq $FILE
+    rm -f $FILE
 fi
-
-# unzip app
-unzip -qq signed.zip
-rm -f unsigned.zip signed.zip
